requirements.txt
asgiref==3.2.3
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
gunicorn==20.1.0
nano==0.10.0
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2==2.9.1
PyJWT==1.7.1
python-decouple==3.5
pytz==2019.3
six==1.16.0
sqlparse==0.3.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
whitenoise==5.3.0
wincertstore==0.2
